refactor: Move Character Creation/Selection segments to character package
