+++ common/src/main/java/com/wynntils/models/character/actionbar/segments/CharacterSelectionSegment.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2025.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.worlds.actionbar.segments;
+package com.wynntils.models.character.actionbar.segments;
 
 import com.wynntils.handlers.actionbar.ActionBarSegment;
